fixing print warning for 664bit
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@11045 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ prcore/tf/include/tf/message_notifier.h
@@ -210,7 +210,7 @@ public:
 
       new_message_queue_.push_back(message);
 
-      NOTIFIER_DEBUG("Added message to message queue, count now %d", new_message_queue_.size());
+      NOTIFIER_DEBUG("Added message to message queue, count now %d", (int)new_message_queue_.size());
 
       new_messages_ = true;
